Added Vector utilities for Exporting. Added Math 1d Stuff. Added Boolean stuff. Android 1d sensors now give a double (always). Android 3d vectors now (Temporarly) give a Vector3d instead of Data3d.

+++ AndroidCodeSnippets/SmaSTraSnippetGenerator/src/main/java/de/tu_darmstadt/smastra/sensors/Abstract1dAndroidSensor.java
@@ -35,7 +35,7 @@ public abstract class Abstract1dAndroidSensor implements SensorEventListener ,de
     /**
      * A cache of the last data.
      */
-    protected float lastData = 0;
+    protected double lastData = 0;
 
 
     /**
@@ -85,7 +85,7 @@ public abstract class Abstract1dAndroidSensor implements SensorEventListener ,de
      * Gets the last Data received.
      * @return the last data.
      */
-    public float getLastData() {
+    public double getLastData() {
         return lastData;
     }
