+++ AndroidCodeSnippets/SmaSTraSnippetGenerator/src/main/java/de/tu_darmstadt/smastra/sensors/AndroidBatteryStatusSensor.java
@@ -37,7 +37,7 @@ public class AndroidBatteryStatusSensor implements Sensor {
      * @return the charger status.
      */
     @SensorOutput
-    public float getStatus(){
+    public double getStatus(){
         IntentFilter ifilter = new IntentFilter(Intent.ACTION_BATTERY_CHANGED);
         Intent batteryStatus = context.registerReceiver(null, ifilter);
         if(batteryStatus == null) return -1;
